% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_controller.R
\name{get_rrq_controller}
\alias{get_rrq_controller}
\title{Try and get an rrq controller}
\usage{
get_rrq_controller(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed through to methods}
}
\description{
Try and get an \code{rrq_controller} object
}

