% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeat_impl.R
\name{rrq_heartbeat}
\alias{rrq_heartbeat}
\title{Create a heartbeat instance}
\description{
Create a heartbeat instance

Create a heartbeat instance
}
\details{
Create a heartbeat instance.  This can be used by running
\code{obj$start()} which will reset the TTL (Time To Live) on \code{key} every
\code{period} seconds (don't set this too high).  If the R process
dies, then the key will expire after \code{3 * period} seconds (or
set \code{expire}) and another application can tell that this R
instance has died.
}
\examples{

if (redux::redis_available()) {
  rand_str <- function() {
    paste(sample(letters, 20, TRUE), collapse = "")
  }
  key <- sprintf("rrq:heartbeat:\%s", rand_str())
  h <- rrq::rrq_heartbeat$new(key, 1, expire = 2)
  con <- redux::hiredis()

  # The heartbeat key exists
  con$EXISTS(key)

  # And has an expiry of less than 2000ms
  con$PTTL(key)

  # We can manually stop the heartbeat, and 2s later the key will
  # stop existing
  h$stop()

  Sys.sleep(2)
  con$EXISTS(key) # 0

  # This is required to close any processes opened by this
  # example, normally you would not need this.
  processx:::supervisor_kill()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rrq_heartbeat-new}{\code{rrq_heartbeat$new()}}
\item \href{#method-rrq_heartbeat-is_running}{\code{rrq_heartbeat$is_running()}}
\item \href{#method-rrq_heartbeat-start}{\code{rrq_heartbeat$start()}}
\item \href{#method-rrq_heartbeat-stop}{\code{rrq_heartbeat$stop()}}
\item \href{#method-rrq_heartbeat-format}{\code{rrq_heartbeat$format()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_heartbeat-new"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_heartbeat-new}{}}}
\subsection{Method \code{new()}}{
Create a heartbeat object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_heartbeat$new(
  key,
  period,
  expire = 3 * period,
  value = expire,
  config = NULL,
  start = TRUE,
  timeout = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Key to use. Once the heartbeat starts it will
create this key and set it to expire after \code{expiry} seconds.}

\item{\code{period}}{Timeout period (in seconds)}

\item{\code{expire}}{Key expiry time (in seconds)}

\item{\code{value}}{Value to store in the key.  By default it stores the
expiry time, so the time since last heartbeat can be computed.
This will be converted to character with \code{as.character}
before saving into Redis}

\item{\code{config}}{Configuration parameters passed through to
\code{redux::redis_config}.  Provide as either a named list or a
\code{redis_config} object.  This allows host, port, password,
db, etc all to be set.}

\item{\code{start}}{Should the heartbeat be started immediately?}

\item{\code{timeout}}{Time, in seconds, to wait for the heartbeat to
appear.  It should generally appear very quickly (within a
second unless your connection is very slow) so this can be
generally left alone.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_heartbeat-is_running"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_heartbeat-is_running}{}}}
\subsection{Method \code{is_running()}}{
Report if heartbeat process is running. This will be
\code{TRUE} if the process has been started and has not stopped.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_heartbeat$is_running()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_heartbeat-start"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_heartbeat-start}{}}}
\subsection{Method \code{start()}}{
Start the heartbeat process. An error will be thrown
if it is already running.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_heartbeat$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_heartbeat-stop"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_heartbeat-stop}{}}}
\subsection{Method \code{stop()}}{
Stop the heartbeat process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_heartbeat$stop(wait = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wait}}{Logical, indicating if we should wait until the
heartbeat process terminates (should take only a fraction of a
second)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_heartbeat-format"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_heartbeat-format}{}}}
\subsection{Method \code{format()}}{
Format method, used by R6 to nicely print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_heartbeat$format(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments, currently ignored}
}
\if{html}{\out{</div>}}
}
}
}
