% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{rrq_default_controller_set}
\alias{rrq_default_controller_set}
\alias{rrq_default_controller_clear}
\title{Register default controller}
\usage{
rrq_default_controller_set(controller)

rrq_default_controller_clear()
}
\arguments{
\item{controller}{An rrq_controller object, or \code{NULL} to clear the
default controller (equivalent to using
\code{rrq_default_controller_clear})}
}
\value{
Invisibly, the previously set default controller (or
\code{NULL} if none was set)
}
\description{
Set or clear a default controller for use with rrq functions.  You
will want to use this to avoid passing \code{controller} in as a named
argument to every function.
}
