% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_misc.R
\name{rrq_deferred_list}
\alias{rrq_deferred_list}
\title{List deferred tasks}
\usage{
rrq_deferred_list(controller = NULL)
}
\arguments{
\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\description{
Return deferred tasks and what they are waiting on.
Note this is in an arbitrary order, tasks will be added to the
queue as their dependencies are satisfied.
}
