% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_misc.R
\name{rrq_destroy}
\alias{rrq_destroy}
\title{Destroy queue}
\usage{
rrq_destroy(
  delete = TRUE,
  worker_stop_type = "message",
  timeout_worker_stop = 0,
  controller = NULL
)
}
\arguments{
\item{delete}{Either \code{TRUE} (the default) indicating that the
keys should be immediately deleted. Alternatively, provide an
integer value and the keys will instead be marked for future
deletion by "expiring" after this many seconds, using Redis'
\code{EXPIRE} command.}

\item{worker_stop_type}{Passed to \verb{$worker_stop}; Can be one of
"message", "kill" or "kill_local". The "kill" method requires that
the workers are using a heartbeat, and "kill_local" requires that
the workers are on the same machine as the controller. However,
these may be faster to stop workers than "message", which will
wait until any task is finished.}

\item{timeout_worker_stop}{A timeout to pass to the worker to
respond the request to stop. See \code{worker_stop}'s \code{timeout}
argument for details.}

\item{controller}{The controller to destroy}
}
\description{
Entirely destroy a queue, by deleting all keys associated with it
from the Redis database. This is a very destructive action and
cannot be undone.
}
