% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir.R
\name{rrq_envir}
\alias{rrq_envir}
\title{Create simple worker environments}
\usage{
rrq_envir(packages = NULL, sources = NULL)
}
\arguments{
\item{packages}{An optional character vector of}

\item{sources}{An optional character vector of scripts to
read. Typically these will contain just function definitions but
you might read large data objects here too.}
}
\value{
A function suitable for passing to the \verb{$envir()} method
of \code{\link[=rrq_controller]{rrq_controller()}}
}
\description{
Helper function for creating a worker environment. This function
exists to create a function suitable for passing to the \code{envir}
method of \code{\link[=rrq_controller]{rrq_controller()}} for the common case where the
worker should source some R scripts and/or load some packages on
startup. This is a convenience wrapper around defining your own
function for use with the \code{envir} method of
\code{\link[=rrq_controller]{rrq_controller()}}, which covers the most simple case. If you
need more flexibility you should write your own.
}
