% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_spawn.R
\name{rrq_expect_workers}
\alias{rrq_expect_workers}
\title{Registed expected workers}
\usage{
rrq_expect_workers(obj, names)
}
\arguments{
\item{obj}{A rrq_controller object}

\item{names}{Names of expected workers}
}
\description{
Register that workers are expected
}

