% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{rrq_key_worker_alive}
\alias{rrq_key_worker_alive}
\title{Randomly generate key for polling new workers}
\usage{
rrq_key_worker_alive(queue_name)
}
\arguments{
\item{queue_name}{Name of the queue (the context id)}
}
\description{
Randomly generate key for polling new workers
}
\author{
Rich FitzJohn
}

