% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{rrq_known_controllers}
\alias{rrq_known_controllers}
\title{Read known controllers}
\usage{
rrq_known_controllers(con = redux::hiredis())
}
\arguments{
\item{con}{A redis connection. The default tries to create a redis
connection using default ports, or environment variables set as in
\code{\link[redux:hiredis]{redux::hiredis()}}}
}
\value{
A \code{data.frame} with columns:
\itemize{
\item \code{id}: the controller identifier
\item \code{time}: the time of last access
\item \code{hostname}: the hostname that last accessed
\item \code{username}: the username that last accessed
}

The \code{data.frame} will be ordered by time of last access, with the
most recenty accesses first.
}
\description{
Read known controllers from the redis database.
}
