% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_message.R
\name{rrq_message_has_response}
\alias{rrq_message_has_response}
\title{Detect if message has response}
\usage{
rrq_message_has_response(
  message_id,
  worker_ids = NULL,
  named = TRUE,
  controller = NULL
)
}
\arguments{
\item{message_id}{The message id}

\item{worker_ids}{Optional vector of worker ids. If \code{NULL} then
all active workers are used (note that this may differ to the set
of workers that the message was sent to!)}

\item{named}{Logical, indicating if the return vector should be named}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A logical vector, possibly named (depending on the \code{named}
argument)
}
\description{
Detect if a response is available for a message
}
