% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_message.R
\name{rrq_message_response_ids}
\alias{rrq_message_response_ids}
\title{Return ids for messages with responses for a particular worker.}
\usage{
rrq_message_response_ids(worker_id, controller = NULL)
}
\arguments{
\item{worker_id}{The worker id}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A character vector of ids
}
\description{
Return ids for messages with responses for a particular worker.
}
