% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_message.R
\name{rrq_message_send}
\alias{rrq_message_send}
\title{Send message to workers}
\usage{
rrq_message_send(command, args = NULL, worker_ids = NULL, controller = NULL)
}
\arguments{
\item{command}{A command, such as \code{PING}, \code{PAUSE}; see the Messages
section of the Details for al messages.}

\item{args}{Arguments to the command, if supported}

\item{worker_ids}{Optional vector of worker ids to send the message
to. If \code{NULL} then the message will be sent to all active workers.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
Invisibly, a single identifier
}
\description{
Send a message to workers. Sending a message returns a message id,
which can be used to poll for a response with the other
\verb{rrq_message_*} functions.  See \code{vignette("messages")} for details
for the messaging interface.
}
\examples{
\dontshow{if (rrq:::enable_examples(require_queue = "rrq:example")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
obj <- rrq_controller("rrq:example")

id <- rrq_message_send("PING", controller = obj)
rrq_message_get_response(id, timeout = 5, controller = obj)
\dontshow{\}) # examplesIf}
}
