% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_misc.R
\name{rrq_queue_length}
\alias{rrq_queue_length}
\title{Queue length}
\usage{
rrq_queue_length(queue = NULL, controller = NULL)
}
\arguments{
\item{queue}{The name of the queue to query (defaults to the
"default" queue).}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A number
}
\description{
Returns the length of the queue (the number of tasks waiting to
run).  This is the same as the length of the value returned by
\link{rrq_queue_list}.
}
