% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_misc.R
\name{rrq_queue_list}
\alias{rrq_queue_list}
\title{List queue contents}
\usage{
rrq_queue_list(queue = NULL, controller = NULL)
}
\arguments{
\item{queue}{The name of the queue to query (defaults to the
"default" queue).}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\description{
Returns the keys in the task queue.
}
