% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_misc.R
\name{rrq_queue_remove}
\alias{rrq_queue_remove}
\title{Remove task ids from a queue}
\usage{
rrq_queue_remove(task_ids, queue = NULL, controller = NULL)
}
\arguments{
\item{task_ids}{Task ids to remove}

\item{queue}{The name of the queue to query (defaults to the
"default" queue).}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\description{
Remove task ids from a queue.
}
