% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_create.R
\name{rrq_task_create_bulk_expr}
\alias{rrq_task_create_bulk_expr}
\title{Create bulk tasks from an expression}
\usage{
rrq_task_create_bulk_expr(
  expr,
  data,
  queue = NULL,
  separate_process = FALSE,
  timeout_task_run = NULL,
  depends_on = NULL,
  controller = NULL
)
}
\arguments{
\item{expr}{An expression, as for \link{rrq_task_create_expr}}

\item{data}{Data that you wish to inject \emph{row-wise} into the expression}

\item{queue}{The queue to add the task to; if not specified the
"default" queue (which all workers listen to) will be used. If
you have configured workers to listen to more than one queue you
can specify that here. Be warned that if you push jobs onto a
queue with no worker, it will queue forever.}

\item{separate_process}{Logical, indicating if the task should be
run in a separate process on the worker. If \code{TRUE}, then the
worker runs the task in a separate process using the \code{callr}
package. This means that the worker environment is completely
clean, subsequent runs are not affected by preceding ones.  The
downside of this approach is a considerable overhead in starting
the external process and transferring data back.}

\item{timeout_task_run}{Optionally, a maximum allowed running
time, in seconds. This parameter only has an effect if
\code{separate_process} is \code{TRUE}. If given, then if the task takes
longer than this time it will be stopped and the task status set
to \code{TIMEOUT}.}

\item{depends_on}{Vector or list of IDs of tasks which must have
completed before this job can be run. Once all dependent tasks
have been successfully run, this task will get added to the
queue. If the dependent task fails then this task will be
removed from the queue.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A character vector with task identifiers; this will have a
length equal to the number of row in \code{data}
}
\description{
Create a bulk set of tasks. Variables in \code{data} take precedence
over variables in the environment in which \code{expr} was
created. There is no "pronoun" support yet (see rlang docs).  Use
\verb{!!} to pull a variable from the environment if you need to, but
be careful not to inject something really large (e.g., any vector
really) or you'll end up with a revolting expression and poor
backtraces.
}
