% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_data}
\alias{rrq_task_data}
\title{Fetch internal task data}
\usage{
rrq_task_data(task_id, controller = NULL)
}
\arguments{
\item{task_id}{A single task identifier}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
Internal data, structures subject to change
}
\description{
Fetch internal data about a task (expert use only)
}
