% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_exists}
\alias{rrq_task_exists}
\title{Test if tasks exist}
\usage{
rrq_task_exists(task_ids, controller = NULL)
}
\arguments{
\item{task_ids}{Vector of task ids to check}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A logical vector the same length as task_ids; \code{TRUE} where
the task exists, \code{FALSE} otherwise.
}
\description{
Test if task ids exist (i.e., are known to this controller).
Nonexistant tasks may be deleted, known to a different controller
or just never have existed.
}
