% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_info}
\alias{rrq_task_info}
\title{Fetch task information}
\usage{
rrq_task_info(task_id, controller = NULL)
}
\arguments{
\item{task_id}{A single task identifier}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A list, format currently subject to change
}
\description{
Fetch information about a task. This currently includes
information about where a task is (or was) running and information
about any retry chain, but will expand in future. The format of
the output here is subject to change (and will probably get a nice
print method) but the values present in the output will be
included in any future update.
}
