% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_log}
\alias{rrq_task_log}
\title{Fetch task logs}
\usage{
rrq_task_log(task_id, controller = NULL)
}
\arguments{
\item{task_id}{A single task identifier}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A character vector of logs, or \code{NULL} if no log is present
yet.  If logging is not enabled for this task, we throw an
error.  Empty logs can be distinguished from "no logs yet", as
they will return an empty character vector (\code{character(0)}).
}
\description{
Fetch logs from tasks that were queued into separate processes
(e.g., with \link{rrq_task_create_expr} using \code{separate_process = TRUE}).  It is not knowable if a task definitely produces logs - if
you have a mixture of workers that do enable task logs and some
that don't, then it will depend on the worker that picks it up if
logging will be enabled.  Don't do this though and you should be
fine.
}
\examples{
\dontshow{if (rrq:::enable_examples(require_queue = "rrq:example")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

obj <- rrq_controller("rrq:example")

t <- rrq_task_create_expr(message("hello!"), separate_process = TRUE,
                          controller = obj)
rrq_task_wait(t, controller = obj)
rrq_task_log(t, controller = obj)
\dontshow{\}) # examplesIf}
}
