% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_position}
\alias{rrq_task_position}
\title{Find task position in queue}
\usage{
rrq_task_position(
  task_ids,
  missing = 0L,
  queue = NULL,
  follow = NULL,
  controller = NULL
)
}
\arguments{
\item{task_ids}{Character vector of tasks to find the position for.}

\item{missing}{Value to return if the task is not found in the queue.
A task will take value \code{missing} if it is running, complete,
errored, deferred etc and a positive integer if it is in the queue,
indicating its position (with 1) being the next task to run.}

\item{queue}{The name of the queue to query (defaults to the
"default" queue).}

\item{follow}{Optional logical, indicating if we should follow any
redirects set up by doing \link{rrq_task_retry}. If not given, falls
back on the value passed into the controller, the global option
\code{rrq.follow}, and finally \code{TRUE}. Set to \code{FALSE} if you want to
return information about the original task, even if it has been
subsequently retried.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
An integer vector, the same length as \code{task_ids}
}
\description{
Find the position of one or more tasks in the queue.
}
