% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_preceeding}
\alias{rrq_task_preceeding}
\title{List tasks ahead of a task}
\usage{
rrq_task_preceeding(task_id, queue = NULL, follow = NULL, controller = NULL)
}
\arguments{
\item{task_id}{Task to find the position for.}

\item{queue}{The name of the queue to query (defaults to the
"default" queue).}

\item{follow}{Optional logical, indicating if we should follow any
redirects set up by doing \link{rrq_task_retry}. If not given, falls
back on the value passed into the controller, the global option
\code{rrq.follow}, and finally \code{TRUE}. Set to \code{FALSE} if you want to
return information about the original task, even if it has been
subsequently retried.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\description{
List the tasks in front of \code{task_id} in the queue.
If the task is missing from the queue this will return NULL. If
the task is next in the queue this will return an empty character
vector.
}
