% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_progress}
\alias{rrq_task_progress}
\title{Fetch task progress information}
\usage{
rrq_task_progress(task_id, controller = NULL)
}
\arguments{
\item{task_id}{A single task id for which the progress is wanted.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
Any set progress object
}
\description{
Retrieve task progress, if set. This will be \code{NULL} if progress
has never been registered, otherwise whatever value was set - can
be an arbitrary R object.
}
