% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_result}
\alias{rrq_task_result}
\title{Fetch single task result}
\usage{
rrq_task_result(task_id, error = FALSE, follow = NULL, controller = NULL)
}
\arguments{
\item{task_id}{The single id for which the result is wanted.}

\item{error}{Logical, indicating if we should throw an error if a
task was not successful. By default (\code{error = FALSE}), in the
case of the task result returning an error we return an object
of class \code{rrq_task_error}, which contains information about the
error. Passing \code{error = TRUE} calls \code{stop()} on this error if it
is returned.}

\item{follow}{Optional logical, indicating if we should follow any
redirects set up by doing \link{rrq_task_retry}. If not given, falls
back on the value passed into the controller, the global option
\code{rrq.follow}, and finally \code{TRUE}. Set to \code{FALSE} if you want to
return information about the original task, even if it has been
subsequently retried.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
The result of your task
}
\description{
Get the result for a single task (see \link{rrq_task_results} for a
method for efficiently getting multiple results at once).  Returns
the value of running the task if it is complete, and an error
otherwise.
}
