% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_results}
\alias{rrq_task_results}
\title{Get the results of a group of tasks, returning them as a list.
See \link{rrq_task_result} for getting the result of a single task.}
\usage{
rrq_task_results(
  task_ids,
  error = FALSE,
  named = FALSE,
  follow = NULL,
  controller = NULL
)
}
\arguments{
\item{task_ids}{A vector of task ids for which the task result
is wanted.}

\item{error}{Logical, indicating if we should throw an error if
the task was not successful. See \code{\link[=rrq_task_result]{rrq_task_result()}} for details.}

\item{named}{Logical, indicating if the return value should be
named with the task ids; as these are quite long this can make
the value a little awkward to work with.}

\item{follow}{Optional logical, indicating if we should follow any
redirects set up by doing \link{rrq_task_retry}. If not given, falls
back on the value passed into the controller, the global option
\code{rrq.follow}, and finally \code{TRUE}. Set to \code{FALSE} if you want to
return information about the original task, even if it has been
subsequently retried.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
An list, one entry per result.  This function errors if
any task is not available.  If \code{named = TRUE}, then this list is
named with the \code{task_ids}.
}
\description{
Get the results of a group of tasks, returning them as a list.
See \link{rrq_task_result} for getting the result of a single task.
}
\examples{
\dontshow{if (rrq:::enable_examples(require_queue = "rrq:example")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
obj <- rrq_controller("rrq:example")

ts <- rrq_task_create_bulk_call(sqrt, 1:10, controller = obj)
rrq_task_wait(ts, controller = obj)
rrq_task_results(ts, controller = obj)

# For a single task, rrq_task_result and rrq_task_results differ
# in the return type; rrq_task_results always returns a list:
t <- ts[[1]]
rrq_task_result(t, controller = obj)
rrq_task_results(t, controller = obj)
\dontshow{\}) # examplesIf}
}
