% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{rrq_worker}
\alias{rrq_worker}
\title{rrq queue worker}
\usage{
rrq_worker(context, con, key_alive = NULL)
}
\arguments{
\item{context}{A context handle object.  The context needs to be loaded.}

\item{con}{A redis connection (redux object).}

\item{key_alive}{Optional key that will be written once the
worker is alive.}
}
\description{
A rrq queue worker.  These are not for interacting with but will
sit and poll a queue for jobs.
}

