% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_config_read}
\alias{rrq_worker_config_read}
\title{Read worker configuration}
\usage{
rrq_worker_config_read(name, timeout = 0, controller = NULL)
}
\arguments{
\item{name}{Name of the configuration (see
\code{\link[=rrq_worker_config_list]{rrq_worker_config_list()}})}

\item{timeout}{Optionally, a timeout to wait for a worker
configuration to appear.  Generally you won't want to set this,
but it can be used to block until a configuration becomes
available.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\description{
Return the value of a of worker configuration saved by
\code{\link[=rrq_worker_config_save]{rrq_worker_config_save()}}
}
\examples{
\dontshow{if (rrq:::enable_examples(require_queue = "rrq:example")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
obj <- rrq_controller("rrq:example")

cfg <- rrq_worker_config("fast")
rrq_worker_config_save("use-fast", cfg, controller = obj)
rrq_worker_config_read("use-fast", controller = obj)
\dontshow{\}) # examplesIf}
}
