% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_config.R
\name{rrq_worker_config_save}
\alias{rrq_worker_config_save}
\title{Save a worker configuration}
\usage{
rrq_worker_config_save(
  queue_id,
  name,
  config,
  overwrite = TRUE,
  con = redux::hiredis()
)
}
\arguments{
\item{queue_id}{The id for the queue}

\item{name}{Name for this configuration}

\item{config}{A worker configuration, created by
\code{\link[=rrq_worker_config]{rrq_worker_config()}}}

\item{overwrite}{Logical, indicating if an existing configuration
with this \code{name} should be overwritten if it exists. If \code{FALSE},
then the configuration is not updated, even if it differs from
the version currently saved.}

\item{con}{The redis connection to use}
}
\value{
Invisibly, a boolean indicating if the configuration was written
}
\description{
Save a worker configuration. This is an alternative (but
equivalent) to setting the worker configuration via the
\verb{$worker_config_save()} method in \link{rrq_controller}, but does
not require setting up a controller (in fact, one need never have
existed).
}
