% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_delete_exited}
\alias{rrq_worker_delete_exited}
\title{Clean up exited workers}
\usage{
rrq_worker_delete_exited(worker_ids = NULL, controller = NULL)
}
\arguments{
\item{worker_ids}{Optional vector of worker ids. If \code{NULL} then
rrq looks for exited workers using \code{\link[=rrq_worker_list_exited]{rrq_worker_list_exited()}}.
If given, we check that the workers are known and have exited.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A character vector of workers that were deleted
}
\description{
Cleans up workers known to have exited
}
