% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{rrq_worker_from_config}
\alias{rrq_worker_from_config}
\title{Create a worker}
\usage{
rrq_worker_from_config(
  queue_id,
  worker_config = "localhost",
  worker_id = NULL,
  con = redux::hiredis(),
  timeout = 5
)
}
\arguments{
\item{queue_id}{The name of the queue}

\item{worker_config}{Optional name of the configuration. The
default "localhost" configuration always exists.}

\item{worker_id}{The id of the worker (if not given a random
id is used)}

\item{con}{Redis configuration}

\item{timeout}{How long to try and read the worker config for. Will
attempt to read once a second and throw an error if config cannot
be located after \code{timeout} seconds.}
}
\value{
A \code{\link{rrq_worker}} object. If you want to "run" the
worker, you would want to use the \verb{$loop()} method of this
object. Other methods are for advanced use.
}
\description{
Create a worker based on a named configuration. This is a wrapper
around directly constructing a \code{\link{rrq_worker}}. Create new
configurations by running \verb{$worker_config_save} from the
\code{\link{rrq_controller}} object.
}
