% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_info}
\alias{rrq_worker_info}
\title{Worker information}
\usage{
rrq_worker_info(worker_ids = NULL, controller = NULL)
}
\arguments{
\item{worker_ids}{Optional vector of worker ids. If \code{NULL} then
all active workers are used.}
}
\value{
A list of \code{worker_info} objects
}
\description{
Returns a list of information about active workers (or exited
workers if \code{worker_ids} includes them).
}
