% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_list_exited}
\alias{rrq_worker_list_exited}
\title{List exited workers}
\usage{
rrq_worker_list_exited(controller = NULL)
}
\arguments{
\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A character vector of worker names
}
\description{
Returns the ids of workers known to have exited
}
