% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_log_tail}
\alias{rrq_worker_log_tail}
\title{Returns the last (few) elements in the worker log, in a
programmatically useful format (see Value).}
\usage{
rrq_worker_log_tail(worker_ids = NULL, n = 1, controller = NULL)
}
\arguments{
\item{worker_ids}{Optional vector of worker ids. If \code{NULL} then
all active workers are used.}

\item{n}{Number of elements to select, the default being the single
last entry. Use \code{Inf} or \code{0} to indicate that you want all log entries}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A \link{data.frame} with columns:
\itemize{
\item \code{worker_id}: the worker id
\item \code{child}: the process id, an integer, where logs come from a child
process from a task queued with \code{separate_process = TRUE}
\item \code{time}: the time from Redis when the event happened; see
\link[redux:redis_time]{redux::redis_time} to convert this to an R time
\item \code{command}: the command sent from or to the worker
\item \code{message}: the message corresponding to that command
}
}
\description{
Returns the last (few) elements in the worker log, in a
programmatically useful format (see Value).
}
\examples{
\dontshow{if (rrq:::enable_examples(require_queue = "rrq:example")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
obj <- rrq_controller("rrq:example")
rrq_worker_log_tail(n = 10, controller = obj)
\dontshow{\}) # examplesIf}
}
