% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_process_log}
\alias{rrq_worker_process_log}
\title{Read worker process log}
\usage{
rrq_worker_process_log(worker_id, controller = NULL)
}
\arguments{
\item{worker_id}{The worker id for which the log is required}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A character vector, one line per line in the log.  If
logging is enabled but the worker has not produced any logs,
this will be an empty character vector. If logging is not
enabled, then this function will throw.
}
\description{
Return the contents of a worker's process log, if
it is located on the same physical storage (including network
storage) as the controller. This will generally behave for
workers started with \link{rrq_worker_spawn} but may require significant
care otherwise.
}
\examples{
\dontshow{if (rrq:::enable_examples(require_queue = "rrq:example")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
obj <- rrq_controller("rrq:example")
worker_id <- rrq_worker_list(controller = obj)[[1]]
tryCatch(
  rrq_worker_process_log(worker_id, controller = obj),
  error = identity)
\dontshow{\}) # examplesIf}
}
