% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_runner.R
\name{rrq_worker_script}
\alias{rrq_worker_script}
\title{Write worker runner script}
\usage{
rrq_worker_script(path, versioned = FALSE)
}
\arguments{
\item{path}{The path to write to. Should be a directory (or one
will be created if it does not yet exist). The final script will
be \code{file.path(path, "rrq_worker")}}

\item{versioned}{Logical, indicating if we should write a
versioned R script that will use the same path to \code{Rscript} as
the running session. If \code{FALSE} we use \verb{#!/usr/bin/env Rscript}
which will pick up \code{Rscript} from the path. You may want to use
a versioned script in tests or if you have multiple R versions
installed simultaneously.}
}
\value{
Invisibly, the path to the script
}
\description{
Write a small script that can be used to launch a rrq worker. The
resulting script takes the same arguments as
\code{\link[=rrq_worker_from_config]{rrq_worker_from_config()}}, but from the command line. See
Details.
}
\details{
If you need to launch rrq workers from a script, it's convenient
not to have to embed R code like:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Rscript -e 'rrq::worker_from_config("myqueue")'
}\if{html}{\out{</div>}}

as this is error-prone and unpleasant to read. You can use the
function \code{rrq_worker_script} to write out a small helper script
which lets you write:

\if{html}{\out{<div class="sourceCode">}}\preformatted{./path/rrq_worker myqueue
}\if{html}{\out{</div>}}

instead.

The helper script supports the same arguments as
\verb{[rrq::rrq_worker_from_config()]}:
\itemize{
\item \code{queue_id} as the sole positional argument
\item \code{worker_config} as \code{--config}
\item \code{worker_id} as \code{--worker-id}
}

To change the redis connection settings, set the \code{REDIS_URL}
environment variable (see \code{\link[redux:hiredis]{redux::hiredis()}} for details).

For example to create a worker \code{myworker} with configuration
\code{myconfig} on queue \code{myqueue} you might use

\if{html}{\out{<div class="sourceCode">}}\preformatted{./rrq_worker --config=myconfig --worker-id=myworker myqueue
}\if{html}{\out{</div>}}
}
\examples{
path <- rrq::rrq_worker_script(tempfile())
readLines(path)
}
