% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_task_id}
\alias{rrq_worker_task_id}
\title{Current task id for workers}
\usage{
rrq_worker_task_id(worker_ids = NULL, controller = NULL)
}
\arguments{
\item{worker_ids}{Optional vector of worker ids. If \code{NULL} then
all active workers are used.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A character vector, \code{NA} where nothing is being worked on,
otherwise corresponding to a task id.
}
\description{
Returns the task id that each worker is working on, if any.
}
