% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_controller.R
\name{worker_controller}
\alias{worker_controller}
\title{Create a worker controller}
\usage{
worker_controller(context_id, con = redux::hiredis())
}
\arguments{
\item{context_id}{A context ID; this will not be checked for
existance, and no root is needed (everything this object does
affects only the Redis database).}

\item{con}{A Redis connection (optional; the default connection
will be used otherwise).}
}
\description{
Create a "worker controller" object, which can deal with
coordinating rrq workers.  This is subject to change.
}

