% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_store.R
\name{object_store_offload_disk}
\alias{object_store_offload_disk}
\title{Disk-based offload}
\description{
Disk-based offload

Disk-based offload
}
\details{
A disk-based offload for \code{\link{object_store}}. This is not
intended at all for direct user-use.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-object_store_offload_disk-new}{\code{object_store_offload_disk$new()}}
\item \href{#method-object_store_offload_disk-mset}{\code{object_store_offload_disk$mset()}}
\item \href{#method-object_store_offload_disk-mget}{\code{object_store_offload_disk$mget()}}
\item \href{#method-object_store_offload_disk-mdel}{\code{object_store_offload_disk$mdel()}}
\item \href{#method-object_store_offload_disk-list}{\code{object_store_offload_disk$list()}}
\item \href{#method-object_store_offload_disk-destroy}{\code{object_store_offload_disk$destroy()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_store_offload_disk-new"></a>}}
\if{latex}{\out{\hypertarget{method-object_store_offload_disk-new}{}}}
\subsection{Method \code{new()}}{
Create the store
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_offload_disk$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{A directory name to store objects in. It will be
created if it does not yet exist.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_store_offload_disk-mset"></a>}}
\if{latex}{\out{\hypertarget{method-object_store_offload_disk-mset}{}}}
\subsection{Method \code{mset()}}{
Save a number of values to disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_offload_disk$mset(hash, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{A character vector of object hashes}

\item{\code{value}}{A list of serialised objects
(each of which is a raw vector)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_store_offload_disk-mget"></a>}}
\if{latex}{\out{\hypertarget{method-object_store_offload_disk-mget}{}}}
\subsection{Method \code{mget()}}{
Retrieve a number of objects from the store
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_offload_disk$mget(hash)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{A character vector of hashes of the objects to return.
The objects will be deserialised before return.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_store_offload_disk-mdel"></a>}}
\if{latex}{\out{\hypertarget{method-object_store_offload_disk-mdel}{}}}
\subsection{Method \code{mdel()}}{
Delete a number of objects from the store
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_offload_disk$mdel(hash)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{A character vector of hashes to remove}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_store_offload_disk-list"></a>}}
\if{latex}{\out{\hypertarget{method-object_store_offload_disk-list}{}}}
\subsection{Method \code{list()}}{
List hashes stored in this offload store
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_offload_disk$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_store_offload_disk-destroy"></a>}}
\if{latex}{\out{\hypertarget{method-object_store_offload_disk-destroy}{}}}
\subsection{Method \code{destroy()}}{
Completely delete the store (by deleting the directory)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_offload_disk$destroy()}\if{html}{\out{</div>}}
}

}
}
