% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_controller.R
\name{rrq_controller}
\alias{rrq_controller}
\title{rrq queue controller}
\usage{
rrq_controller(queue_id, con = redux::hiredis())
}
\arguments{
\item{queue_id}{An identifier for the queue.  This will prefix all
keys in redis, so a prefix might be useful here depending on
your use case (e.g. \code{rrq:<user>:<id>})}

\item{con}{A redis connection}
}
\description{
A queue controller.  Use this to interact with a queue/cluster.
}
