% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_controller.R
\name{rrq_controller}
\alias{rrq_controller}
\title{rrq queue controller}
\usage{
rrq_controller(context, con)
}
\arguments{
\item{context}{A context handle object.  The context needs to be loaded.}

\item{con}{A redis connection (redux object).}
}
\description{
A queue controller.  Use this to interact with a queue/cluster.
}

