% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{rrq_default_controller_set}
\alias{rrq_default_controller_set}
\alias{rrq_default_controller_clear}
\title{Register default controller}
\usage{
rrq_default_controller_set(controller)

rrq_default_controller_clear()
}
\arguments{
\item{controller}{An rrq_controller2 object}
}
\description{
Set or clear a default controller for use with rrq functions.  You
will want to use this to avoid passing \code{controller} in as a named
argument to every function.
}
