% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_spawn.R
\name{rrq_expect_worker}
\alias{rrq_expect_worker}
\title{Register expected workers}
\usage{
rrq_expect_worker(obj, ids)
}
\arguments{
\item{obj}{A rrq_controller object}

\item{ids}{Ids of expected workers}
}
\description{
Register that workers are expected.  This generates a key that one
or more workers will write to when they start up (as used by
\code{rrq_worker_spawn}).
}
