% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_data}
\alias{rrq_task_data}
\title{Fetch internal task data}
\usage{
rrq_task_data(task_id, controller = NULL)
}
\arguments{
\item{task_id}{A single task identifier}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
Internal data, structures subject to change
}
\description{
Fetch internal data about a task (expert use only)
}
\examples{
\dontshow{if (rrq:::enable_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

obj <- rrq_controller("rrq:example")

t <- rrq_task_create_expr(runif(1), controller = obj)
rrq_task_data(t, controller = obj)

x <- 10
y <- 20
t <- rrq_task_create_expr(x + y, controller = obj)
rrq_task_data(t, controller = obj)
\dontshow{\}) # examplesIf}
}
