% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_delete}
\alias{rrq_task_delete}
\title{Delete tasks}
\usage{
rrq_task_delete(task_ids, check = TRUE, controller = NULL)
}
\arguments{
\item{task_ids}{Vector of task ids to delete}

\item{check}{Logical indicating if we should check that the tasks
are not running. Deleting running tasks is unlikely to result in
desirable behaviour.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
Nothing, called for side effects only
}
\description{
Delete one or more tasks
}
