% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_exists}
\alias{rrq_task_exists}
\title{Test if tasks exist}
\usage{
rrq_task_exists(task_ids, named = FALSE, controller = NULL)
}
\arguments{
\item{task_ids}{Vector of task ids to check}

\item{named}{Logical, indicating if the return value should be
named with the task ids; as these are quite long this can make
the value a little awkward to work with.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A logical vector the same length as task_ids; \code{TRUE} where
the task exists, \code{FALSE} otherwise.  If \code{named} was \code{TRUE}, then
this vector is named with \code{task_ids}.
}
\description{
Test if task ids exist (i.e., are known to this controller).
Nonexistent tasks may be deleted, known to a different controller
or just never have existed.
}
\examples{
\dontshow{if (rrq:::enable_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
obj <- rrq_controller("rrq:example")

t1 <- rrq_task_create_expr(runif(1), controller = obj)
rrq_task_exists(t1, controller = obj)

t2 <- ids::random_id()
rrq_task_exists(t2, controller = obj)
\dontshow{\}) # examplesIf}
}
