% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_results}
\alias{rrq_task_results}
\title{Get the results of a group of tasks, returning them as a list.
See \link{rrq_task_result} for getting the result of a single task.}
\usage{
rrq_task_results(task_ids, error = FALSE, follow = NULL, controller = NULL)
}
\arguments{
\item{task_ids}{A vector of task ids for which the task result
is wanted.}

\item{error}{Logical, indicating if we should throw an error if
the task was not successful. See \code{\link[=rrq_task_result]{rrq_task_result()}} for details.}

\item{follow}{Optional logical, indicating if we should follow any
redirects set up by doing \link{rrq_task_retry}. If not given, falls
back on the value passed into the controller, the global option
\code{rrq.follow}, and finally \code{TRUE}. Set to \code{FALSE} if you want to
return information about the original task, even if it has been
subsequently retried.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
An unnamed list, one entry per result.  This function
errors if any task is not available.
}
\description{
Get the results of a group of tasks, returning them as a list.
See \link{rrq_task_result} for getting the result of a single task.
}
