% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_task.R
\name{rrq_task_retry}
\alias{rrq_task_retry}
\title{Retry tasks}
\usage{
rrq_task_retry(task_ids, controller = NULL)
}
\arguments{
\item{task_ids}{Task ids to retry.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
New task ids
}
\description{
Retry a task (or set of tasks). Typically this is after failure
(e.g., \code{ERROR}, \code{DIED} or similar) but you can retry even
successfully completed tasks. Once retried, methods that retrieve
information about a task (e.g., \verb{$task_status()},
\verb{$task_result()}) will behave differently depending on the value
of their \code{follow} argument. See \code{vignette("fault-tolerance")} for
more details.
}
