% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{rrq_worker}
\alias{rrq_worker}
\title{rrq queue worker}
\description{
rrq queue worker

rrq queue worker
}
\details{
A rrq queue worker.  These are not typically for interacting with
but will sit and poll a queue for jobs.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the worker}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rrq_worker-new}{\code{rrq_worker$new()}}
\item \href{#method-rrq_worker-info}{\code{rrq_worker$info()}}
\item \href{#method-rrq_worker-log}{\code{rrq_worker$log()}}
\item \href{#method-rrq_worker-load_envir}{\code{rrq_worker$load_envir()}}
\item \href{#method-rrq_worker-poll}{\code{rrq_worker$poll()}}
\item \href{#method-rrq_worker-step}{\code{rrq_worker$step()}}
\item \href{#method-rrq_worker-loop}{\code{rrq_worker$loop()}}
\item \href{#method-rrq_worker-format}{\code{rrq_worker$format()}}
\item \href{#method-rrq_worker-timer_start}{\code{rrq_worker$timer_start()}}
\item \href{#method-rrq_worker-progress}{\code{rrq_worker$progress()}}
\item \href{#method-rrq_worker-task_eval}{\code{rrq_worker$task_eval()}}
\item \href{#method-rrq_worker-shutdown}{\code{rrq_worker$shutdown()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-new"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-new}{}}}
\subsection{Method \code{new()}}{
Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$new(
  queue_id,
  con = redux::hiredis(),
  key_alive = NULL,
  worker_name = NULL,
  queue = NULL,
  time_poll = NULL,
  timeout_idle = NULL,
  heartbeat_period = NULL,
  verbose = TRUE,
  is_child = FALSE,
  timeout_poll = 1,
  timeout_die = 2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{queue_id}}{Name of the queue to connect to.  This will be the
prefix to all the keys in the redis database}

\item{\code{con}}{A redis connection}

\item{\code{key_alive}}{Optional key that will be written once the
worker is alive.}

\item{\code{worker_name}}{Optional worker name.  If omitted, a random
name will be created.}

\item{\code{queue}}{Queues to listen on, listed in decreasing order of
priority. If not given, we listen on the "default" queue.}

\item{\code{time_poll}}{Poll time.  Longer values here will reduce the
impact on the database but make workers less responsive to being
killed with an interrupt (control-C or Escape).  The default
should be good for most uses, but shorter values are used for
debugging.}

\item{\code{timeout_idle}}{Optional timeout to set for the worker.  This is
(roughly) equivalent to issuing a \code{TIMEOUT_SET} message
after initialising the worker, except that it's guaranteed to be
run by all workers.}

\item{\code{heartbeat_period}}{Optional period for the heartbeat.  If
non-NULL then a heartbeat process will be started (using
\code{\link{rrq_heartbeat}}) which can be used to build fault
tolerant queues.}

\item{\code{verbose}}{Logical, indicating if the worker should print
logging output to the screen.  Logging to screen has a small but
measurable performance cost, and if you will not collect system
logs from the worker then it is wasted time.  Logging to the
redis server is always enabled.}

\item{\code{is_child}}{Logical, used to indicate that this is a child of
the real worker.  If \code{is_child} is \code{TRUE}, then most other
arguments here have no effect (e.g., \code{queue} all the timeout /
idle / polling arguments) as they come from the parent.}

\item{\code{timeout_poll}}{Optional timeout indicating how long to wait
for a background process to produce stdout or stderr. Only used
for tasks queued with \code{separate_process} \code{TRUE}.}

\item{\code{timeout_die}}{Optional timeout indicating how long to wait
wait for the background process to respond to SIGTERM before
we stop the worker. Only used for tasks queued with
\code{separate_process} \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-info"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-info}{}}}
\subsection{Method \code{info()}}{
Return information about this worker, a list of
key-value pairs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-log"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-log}{}}}
\subsection{Method \code{log()}}{
Create a log entry. This will print a human readable
format to screen and a machine-readable format to the redis database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$log(label, value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{Scalar character, the title of the log entry}

\item{\code{value}}{Character vector (or null) with log values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-load_envir"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-load_envir}{}}}
\subsection{Method \code{load_envir()}}{
Load the worker environment by creating a new
environment object and running the create hook (if configured).
See \verb{$envir} on \code{\link{rrq_controller}} for details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$load_envir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-poll"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-poll}{}}}
\subsection{Method \code{poll()}}{
Poll for work
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$poll(immediate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{immediate}}{Logical, indicating if we should \emph{not}
do a blocking wait on the queue but instead reducing the timeout to
zero. Intended primarily for use in the tests.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-step"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-step}{}}}
\subsection{Method \code{step()}}{
Take a single "step". This consists of
\enumerate{
\item Poll for work (\verb{$poll()})
\item If work found, run it (either a task or a message)
\item If work not found, check the timeout
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$step(immediate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{immediate}}{Logical, indicating if we should \emph{not}
do a blocking wait on the queue but instead reducing the timeout to
zero. Intended primarily for use in the tests.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-loop"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-loop}{}}}
\subsection{Method \code{loop()}}{
The main worker loop. Use this to set up the main
worker event loop, which will continue until exiting (via a timeout
or message).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$loop(immediate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{immediate}}{Logical, indicating if we should \emph{not}
do a blocking wait on the queue but instead reducing the timeout to
zero. Intended primarily for use in the tests.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-format"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-format}{}}}
\subsection{Method \code{format()}}{
Create a nice string representation of the worker.
Used automatically to print the worker by R6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-timer_start"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-timer_start}{}}}
\subsection{Method \code{timer_start()}}{
Start the timer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$timer_start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-progress"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-progress}{}}}
\subsection{Method \code{progress()}}{
Submit a progress message. See
\code{\link[=rrq_task_progress_update]{rrq_task_progress_update()}} for details of this mechanism.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$progress(value, error = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{An R object with the contents of the update. This
will overwrite any previous progress value, and can be retrieved
from a \link{rrq_controller} with the
\verb{$task_progress} method.  A value of \code{NULL} will appear
to clear the status, as \code{NULL} will also be returned if no
status is found for a task.}

\item{\code{error}}{Logical, indicating if we should throw an error if
not running as an \code{rrq} task. Set this to \code{FALSE} if
you want code to work without modification within and outside of
an \code{rrq} job, or to \code{TRUE} if you want to be sure that
progress messages have made it to the server.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-task_eval"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-task_eval}{}}}
\subsection{Method \code{task_eval()}}{
Evaluate a task
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$task_eval(task_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task_id}}{A task identifier. It is undefined what happens if
this identifier does not exist.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrq_worker-shutdown"></a>}}
\if{latex}{\out{\hypertarget{method-rrq_worker-shutdown}{}}}
\subsection{Method \code{shutdown()}}{
Stop the worker
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rrq_worker$shutdown(status = "OK", graceful = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{the worker status; typically be one of \code{OK} or \code{ERROR}
but can be any string}

\item{\code{graceful}}{Logical, indicating if we should request a
graceful shutdown of the heartbeat, if running.}
}
\if{html}{\out{</div>}}
}
}
}
