% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_config_read}
\alias{rrq_worker_config_read}
\title{Read worker configuration}
\usage{
rrq_worker_config_read(name, controller = NULL)
}
\arguments{
\item{name}{Name of the configuration (see
\code{\link[=rrq_worker_config_list]{rrq_worker_config_list()}})}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\description{
Return the value of a of worker configuration saved by
\code{\link[=rrq_worker_config_save2]{rrq_worker_config_save2()}}
}
