% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_config_save}
\alias{rrq_worker_config_save}
\title{Save worker configuration}
\usage{
rrq_worker_config_save(name, config, overwrite = TRUE, controller = NULL)
}
\arguments{
\item{name}{Name for this configuration}

\item{config}{A worker configuration, created by
\code{\link[=rrq_worker_config]{rrq_worker_config()}}}

\item{overwrite}{Logical, indicating if an existing configuration
with this \code{name} should be overwritten if it exists. If \code{FALSE},
then the configuration is not updated, even if it differs from
the version currently saved.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
Invisibly, a boolean indicating if the configuration was
updated.
}
\description{
Save a worker configuration, which can be used to start workers
with a set of options with the cli. These correspond to arguments
to \link{rrq_worker}. \strong{This function will be renamed soon}
}
