% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_envir_set}
\alias{rrq_worker_envir_set}
\title{Set worker environment}
\usage{
rrq_worker_envir_set(create, notify = TRUE, controller = NULL)
}
\arguments{
\item{create}{A function that will create an environment. It will
be called with one parameter (an environment), in a fresh R
session. The function \code{\link[=rrq_envir]{rrq_envir()}} can be used to
create a suitable function for the most common case (loading
packages and sourcing scripts).}

\item{notify}{Boolean, indicating if we should send a \code{REFRESH}
message to all workers to update their environment.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\description{
Register a function to create an environment when creating a
worker. When a worker starts, they will run this function.
}
