% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_len}
\alias{rrq_worker_len}
\title{Number of active workers}
\usage{
rrq_worker_len(controller = NULL)
}
\arguments{
\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
An integer
}
\description{
Returns the number of active workers
}
