% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_list}
\alias{rrq_worker_list}
\title{List active workers}
\usage{
rrq_worker_list(controller = NULL)
}
\arguments{
\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A character vector of worker names
}
\description{
Returns the ids of active workers.  This does not include exited
workers; use \code{\link[=rrq_worker_list_exited]{rrq_worker_list_exited()}} for that.
}
