% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_process_log}
\alias{rrq_worker_process_log}
\title{Read worker process log}
\usage{
rrq_worker_process_log(worker_id, controller = NULL)
}
\arguments{
\item{worker_id}{The worker id for which the log is required}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\description{
Return the contents of a worker's process log, if
it is located on the same physical storage (including network
storage) as the controller. This will generally behave for
workers started with \link{rrq_worker_spawn} but may require significant
care otherwise.
}
