% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq_workers.R
\name{rrq_worker_status}
\alias{rrq_worker_status}
\title{Worker statuses}
\usage{
rrq_worker_status(worker_ids = NULL, controller = NULL)
}
\arguments{
\item{worker_ids}{Optional vector of worker ids. If \code{NULL} then
all active workers are used.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\value{
A character vector of statuses, named by worker
}
\description{
Returns a character vector of current worker statuses
}
