% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_spawn.R
\name{rrq_worker_wait}
\alias{rrq_worker_wait}
\title{Wait for workers}
\usage{
rrq_worker_wait(
  worker_ids,
  timeout = Inf,
  time_poll = 1,
  progress = NULL,
  controller = NULL
)
}
\arguments{
\item{worker_ids}{A vector of worker ids to wait for}

\item{timeout}{Timeout in seconds; default is to wait forever}

\item{time_poll}{Poll interval, in seconds. Must be an integer}

\item{progress}{Optional logical indicating if a progress bar
should be displayed. If \code{NULL} we fall back on the value of the
global option \code{rrq.progress}, and if that is unset display a
progress bar if in an interactive session.}

\item{controller}{The controller to use.  If not given (or \code{NULL})
we'll use the controller registered with
\code{\link[=rrq_default_controller_set]{rrq_default_controller_set()}}.}
}
\description{
Wait for workers to appear.
}
