% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_spawn.R
\name{worker_spawn}
\alias{worker_spawn}
\alias{worker_wait}
\title{Spawn a worker}
\usage{
worker_spawn(obj, n = 1, logdir = "worker_logs", timeout = 600,
  worker_config = "localhost", worker_name_base = NULL, path = NULL,
  time_poll = 1, progress = NULL)

worker_wait(obj, key_alive, timeout = 600, time_poll = 1, progress = NULL)
}
\arguments{
\item{obj}{An \code{rrq_controller} object}

\item{n}{Number of workers to spawn}

\item{logdir}{Path of a log directory to write to.  This will be
interpreted relative to the context root.}

\item{timeout}{Time to wait for workers to appear}

\item{worker_config}{Name of the configuration to use.  By default
the \code{"localhost"} configuration is used, which duplicates
the redis information from the local configuration but otherwise
uses the defaults.}

\item{worker_name_base}{Optional base to construct the worker
names from.  If omitted a random name will be used.}

\item{path}{Path to start the worker in.  By default workers will
start in the current working directory, but you can start them
elsewhere by providing a path here.  If the path does not exist,
an error will be thrown.  If \code{n} is greater than 1, all
workers will start in the same working directory.  The
\code{logfile} argument will be interpreted relative to current
working directory (not the worker working directory); use
\code{\link{normalizePath}} to convert into an absolute path
name to prevent this.}

\item{time_poll}{Polling period (in seconds) while waiting for
workers to come up.  Must be an integer, at least 1.}

\item{progress}{Show a progress bar while waiting for workers
(when \code{timeout} is at least 0)}

\item{key_alive}{A key name (generated from
\code{\link{rrq_expect_worker}} or \code{worker_spawn})}
}
\description{
Spawn a worker in the background
}
\details{
Spawning multiple workers.  If \code{n} is greater than one,
multiple workers will be spawned.  This happens in parallel so it
does not take n times longer than spawing a single worker.

Beware that signals like Ctrl-C passed to \emph{this} R instance
can still propagate to the child processes and can result in them
dying unexpectedly.  It is probably safer to start processes in a
completely separate session.
}

